requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
